requirements.txt
pybit
requests
python-telegram-bot
python-binance
pymongo